add: Navigatorのインターフェースをfeatureのuiに定義するように変更
