+++ apps/app/lib/router/navigator/home_navigator.dart
@@ -1,6 +1,6 @@
-import 'package:cores_navigation/navigators.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_app/router/provider/router.dart';
+import 'package:flutter_app/ui/home_page.dart';
 
 final class HomeNavigatorImpl implements HomeNavigator {
   const HomeNavigatorImpl();
@@ -13,14 +13,5 @@ final class HomeNavigatorImpl implements HomeNavigator {
   @override
   void goWebPage(BuildContext context) {
     const WebPageRoute().go(context);
-  }
-
-  @override
-  void goGitHubRepositoryDetailPage(
-    BuildContext context,
-    String repositoryName,
-    String? description,
-  ) {
-    GitHubRepositoryDetailPageRoute(repositoryName, description).go(context);
   }
 }
