+++ apps/app/lib/ui/provider/navigator_provider.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_app/ui/home_page.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'navigator_provider.g.dart';
+
+@Riverpod(dependencies: [])
+HomeNavigator homeNavigator(HomeNavigatorRef ref) => throw UnimplementedError();
