--- packages/cores/navigation/.metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: "2e9cb0aa71a386a91f73f7088d115c0d96654829"
-  channel: "stable"
-
-project_type: package
