--- packages/cores/navigation/analysis_options.yaml
@@ -1,6 +1,0 @@
-include: package:yumemi_lints/flutter/3.22/recommended.yaml
-
-analyzer:
-  plugins:
-    # https://riverpod.dev/docs/introduction/getting_started#enabling-riverpod_lintcustom_lint
-    - custom_lint
