--- packages/cores/navigation/lib/navigators.dart
@@ -1,2 +1,0 @@
-export 'src/home/navigator.dart';
-export 'src/setting/navigator.dart';
