--- packages/cores/navigation/lib/providers.dart
@@ -1,2 +1,0 @@
-export 'src/home/provider.dart';
-export 'src/setting/provider.dart';
