--- packages/cores/navigation/lib/src/home/navigator.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-
-abstract interface class HomeNavigator {
-  void goDebugPage(BuildContext context);
-  void goWebPage(BuildContext context);
-  void goGitHubRepositoryDetailPage(
-    BuildContext context,
-    String repositoryName,
-    String? description,
-  );
-}
