--- packages/cores/navigation/lib/src/home/provider.dart
@@ -1,7 +1,0 @@
-import 'package:cores_navigation/src/home/navigator.dart';
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-part 'provider.g.dart';
-
-@riverpod
-external HomeNavigator homeNavigator(HomeNavigatorRef ref);
