--- packages/cores/navigation/lib/src/setting/navigator.dart
@@ -1,5 +1,0 @@
-import 'package:flutter/material.dart';
-
-abstract interface class SettingNavigator {
-  void goLicensePage(BuildContext context);
-}
