--- packages/cores/navigation/lib/src/setting/provider.dart
@@ -1,7 +1,0 @@
-import 'package:cores_navigation/src/setting/navigator.dart';
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-part 'provider.g.dart';
-
-@riverpod
-external SettingNavigator settingNavigator(SettingNavigatorRef ref);
