--- packages/cores/navigation/pubspec.yaml
@@ -1,24 +1,0 @@
-name: cores_navigation
-description: "A new Flutter package project."
-version: 0.0.1
-
-environment:
-  sdk: ^3.4.1
-  flutter: ^3.22.1
-
-dependencies:
-  flutter:
-    sdk: flutter
-  flutter_riverpod: ^2.5.1
-  riverpod_annotation: ^2.3.5
-
-dev_dependencies:
-  build_runner: ^2.4.7
-  custom_lint: ^0.6.4
-  flutter_test:
-    sdk: flutter
-  riverpod_generator: ^2.4.0
-  riverpod_lint: ^2.3.10
-  yumemi_lints: ^2.1.0
-
-flutter:
