+++ packages/features/github_repository/lib/src/ui/github_repository_list.dart
@@ -1,13 +1,21 @@
 import 'package:cores_core/exception.dart';
 import 'package:cores_core/ui.dart';
 import 'package:cores_designsystem/components.dart';
-import 'package:cores_navigation/providers.dart';
 import 'package:features_github_repository/src/data/provider/scroll_notifier.dart';
 import 'package:features_github_repository/src/domain/model/github_repository.dart';
 import 'package:features_github_repository/src/ui/provider/github_repository_list_notifier.dart';
+import 'package:features_github_repository/ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+abstract interface class GithubRepositoryListNavigator {
+  void goGitHubRepositoryDetailPage(
+    BuildContext context,
+    String repositoryName,
+    String? description,
+  );
+}
 
 typedef GitHubRepositoryPagingView = CommonPagingView<
     PageBasedGitHubRepositoryNotifier,
@@ -20,7 +28,7 @@ class GitHubRepositoryList extends HookConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final scrollController = useScrollController();
-    final navigator = ref.watch(homeNavigatorProvider);
+    final navigator = ref.watch(githubRepositoryListNavigatorProvider);
     ref.listen(scrollNotifierProvider, (_, __) async {
       if (!context.mounted) {
         return;
