+++ packages/features/github_repository/lib/src/ui/provider/navigator_provider.dart
@@ -1,0 +1,10 @@
+import 'package:features_github_repository/src/ui/github_repository_list.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'navigator_provider.g.dart';
+
+@Riverpod(dependencies: [])
+GithubRepositoryListNavigator githubRepositoryListNavigator(
+  GithubRepositoryListNavigatorRef ref,
+) =>
+    throw UnimplementedError();
