+++ packages/features/github_repository/lib/ui.dart
@@ -1,2 +1,3 @@
 export 'src/ui/github_repository_detail_page.dart';
 export 'src/ui/github_repository_list.dart';
+export 'src/ui/provider/navigator_provider.dart';
