+++ packages/features/setting/build.yaml
@@ -1,0 +1,27 @@
+targets:
+  $default:
+    builders:
+      freezed:
+        # https://github.com/dart-lang/build/blob/master/docs/faq.md#how-do-i-avoid-running-builders-on-unnecessary-inputs
+        generate_for:
+          include:
+            - lib/{model,**/model}/**.dart
+      # https://github.com/google/json_serializable.dart/tree/master/json_serializable#build-configuration
+      json_serializable:
+        generate_for:
+          include:
+            - lib/{model,**/model}/**.dart
+        options:
+          field_rename: snake
+          # json のデシリアライズ時に発生する Exception を CheckedFromJsonException にまとめる
+          checked: true
+      riverpod_generator:
+        generate_for:
+          include:
+            - lib/{provider,**/provider}/**.dart
+      # https://github.com/dart-lang/source_gen#ignore_for_file
+      source_gen:combining_builder:
+        options:
+          ignore_for_file:
+            - type=lint
+            - duplicate_ignore
