--- packages/features/setting/lib/setting.dart
@@ -1,1 +1,0 @@
-export 'src/ui/setting_page.dart';
