+++ packages/features/setting/lib/src/ui/provider/navigator_provider.dart
@@ -1,0 +1,8 @@
+import 'package:features_setting/src/ui/setting_page.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'navigator_provider.g.dart';
+
+@Riverpod(dependencies: [])
+SettingNavigator settingNavigator(SettingNavigatorRef ref) =>
+    throw UnimplementedError();
