+++ packages/features/setting/lib/src/ui/setting_page.dart
@@ -1,12 +1,16 @@
 import 'package:cores_core/provider.dart';
 import 'package:cores_data/theme_mode.dart';
-import 'package:cores_navigation/providers.dart';
 import 'package:features_setting/src/gen/assets/assets.gen.dart';
 import 'package:features_setting/src/gen/l10n/l10n.dart';
 import 'package:features_setting/src/ui/components/setting_section_spacer.dart';
 import 'package:features_setting/src/ui/components/setting_section_title.dart';
+import 'package:features_setting/src/ui/provider/navigator_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+abstract interface class SettingNavigator {
+  void goLicensePage(BuildContext context);
+}
 
 class SettingPage extends ConsumerWidget {
   const SettingPage({super.key});
