+++ packages/features/setting/lib/ui.dart
@@ -1,0 +1,2 @@
+export 'src/ui/provider/navigator_provider.dart';
+export 'src/ui/setting_page.dart';
