+++ packages/features/setting/pubspec.yaml
@@ -12,18 +12,19 @@ dependencies:
     path: ../../cores/core
   cores_data:
     path: ../../cores/data
-  cores_navigation:
-    path: ../../cores/navigation
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
   flutter_riverpod: ^2.5.1
+  freezed_annotation: ^2.4.1
   intl: ^0.19.0
+  riverpod_annotation: ^2.3.5
 
 dev_dependencies:
   build_runner: ^2.4.7
   flutter_gen_runner: ^5.5.0+1
+  riverpod_generator: ^2.4.0
   yumemi_lints: ^2.1.0
 
 flutter:
